Crate human_request_agent

Human Request Agent

This agent will be the responsible determining which Sre agent need to do which task from human request and reporting to Main Agent

Modules

human_request_node

here we will have nodes human request agent centered this will hold the node logic using prompts specific and structured output schema specific to any mini-node we have broken down the agents task into smaller ones easier to digest and compatible with local llm limitations so ‘conditional edges’ will be in the bigger function that coordinated the nodes